+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - [#3](https://github.com/josephscott/marmerine/pull/3) @joanhey - Call time() instead of microtime() for uptime tracking
 - [#4](https://github.com/josephscott/marmerine/pull/4) @joanhey - Track Marmerine version as a constant
 - [#5](https://github.com/josephscott/marmerine/pull/5) @joanhey - Use static anonymous functions, provides a small memory improvement
+- [#8](https://github.com/josephscott/marmerine/pull/8) @joanhey - Set the workerman name
+- [#7](https://github.com/josephscott/marmerine/pull/7) @joanhey - Remove int casting from verbose()
+- [#9](https://github.com/josephscott/marmerine/pull/9) @joanhey - Test file naming style changes
+- Set a busy timeout for the SQLite database, and always enable WAL
+- [#12](https://github.com/josephscott/marmerine/pull/12) @joanhey - Send ERROR when a command is not supported
 
 
 ## 0.0.3 - 2022-05-21